requirements.txt
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
docx2txt
PyPDF2